+++ CHANGELOG.md
@@ -6,7 +6,15 @@
   <a href="CHANGELOG_EN.md">English</a>
 </p>
 
-## v1.9.0(main)
+## v1.10.0(main)
+
+### Features
+
+### Bug Fixes
+
+### Improvements
+
+## v1.9.0
 
 ### Features